move functions into util
