+++ selfdrive/ui/replay/logreader.cc
@@ -1,43 +1,22 @@
 #include "selfdrive/ui/replay/logreader.h"
 
 #include <cassert>
-#include <bzlib.h>
+#include <sstream>
 #include "selfdrive/common/util.h"
-
-static bool decompressBZ2(std::vector<uint8_t> &dest, const char srcData[], size_t srcSize,
-                          size_t outputSizeIncrement = 0x100000U) {
-  bz_stream strm = {};
-  int ret = BZ2_bzDecompressInit(&strm, 0, 0);
-  assert(ret == BZ_OK);
-
-  strm.next_in = const_cast<char *>(srcData);
-  strm.avail_in = srcSize;
-  do {
-    strm.next_out = (char *)&dest[strm.total_out_lo32];
-    strm.avail_out = dest.size() - strm.total_out_lo32;
-    ret = BZ2_bzDecompress(&strm);
-    if (ret == BZ_OK && strm.avail_in > 0 && strm.avail_out == 0) {
-      dest.resize(dest.size() + outputSizeIncrement);
-    }
-  } while (ret == BZ_OK);
-
-  BZ2_bzDecompressEnd(&strm);
-  dest.resize(strm.total_out_lo32);
-  return ret == BZ_STREAM_END;
-}
+#include "selfdrive/ui/replay/util.h"
 
 LogReader::~LogReader() {
   for (auto e : events) delete e;
 }
 
 bool LogReader::load(const std::string &file) {
-  raw_.resize(1024 * 1024 * 64);
-  std::string dat = util::read_file(file);
-  if (dat.empty() || !decompressBZ2(raw_, dat.data(), dat.size())) {
+  std::stringstream stream;
+  if (!readBZ2File(file, stream)) {
     LOGW("bz2 decompress failed");
     return false;
   }
 
+  raw_ = stream.str();
   kj::ArrayPtr<const capnp::word> words((const capnp::word *)raw_.data(), raw_.size() / sizeof(capnp::word));
   while (words.size() > 0) {
     try {
